新たに ToggleStatusButtonGroup.tsx を作成しました。
HeaderTitle.tsx 内の３つの toggleStatusGroup をまとめました。
